requirements.txt.txt
requests
Flask
render_template
request
url_for
redirect
session
SQLAlchemy 
text
mysql.connector
panas
MySQLdb
json_normalize
